# Improve {u}int/long.ToString/TryFormat throughput by pre-computing the length